requirements.txt
asgiref==3.5.0
Django==3.2.8
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
